requirements.txt
django
pillow
python-dotenv
gunicorn
django-allauth
requests
PyJWT
cryptography